+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+=== Head
+
+* Support array of patterns for `ignoreSelectors`.
+
 === 1.1.0 (September 18, 2015)
 
 * Support selective overriding of a chosen preset's patterns.